+++ src/main/java/lkm/starterproject/accounting/repository/basic/TaxOfficeRepository.java
@@ -3,5 +3,5 @@ package lkm.starterproject.accounting.repository.basic;
 import lkm.starterproject.accounting.entity.basic.TaxOffice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaxOfficeRepository extends JpaRepository<TaxOffice, Integer> {
+public interface TaxOfficeRepository extends JpaRepository<TaxOffice, Long> {
 }
